backend/requirements-voice-minimal.txt
# Minimal Voice Diarization Dependencies for production
# Only includes packages absolutely necessary for TitaNet + AssemblyAI workflow

# Core voice processing
assemblyai>=0.34.0
pydub>=0.25.1

# PyArrow - needed by some dependencies
pyarrow>=12.0.0,<18.0.0

# NeMo core (for TitaNet speaker verification)
nemo-toolkit[asr]==1.21.0

# Essential NeMo dependencies
hydra-core>=1.3.0
pytorch-lightning>=2.0.0
librosa>=0.10.0

# Model downloads
huggingface-hub==0.23.5

# Text processing
sentencepiece>=0.1.99
youtokentome>=1.0.6

# Monitoring and health checks
requests>=2.31.0
psutil>=5.9.0

# Note: PyTorch with CUDA is installed separately in Dockerfile.gpu
# Note: Cython is installed separately in Dockerfile.gpu
# Note: Removed non-essential packages: transformers, pyannote.audio, webdataset, lhotse, datasets, editdistance, jiwer, inflect
